Fixes bug in changing language when viewing the app locally (file:// protocol)
When the app runs in local, jQuery ajax requests do not guess the content type correctly.
For example, when requesting the *_messages.js files, jQuery is not capable of
guessing the content (since no HTTP response headers are available on file://), so
it tries to parse the response as XML. In order to avoid that, the jQuery ajax request
is made by setting the expected response content-type ('text').